refactor: update type imports and remove unused types across components
