+++ src/types/folder/index.ts
@@ -8,8 +8,6 @@ import React from 'react';
 
 import { IncompleteMessage, MailMessage } from 'types/messages';
 
-export type GrantType = { gt: string; perm: string; zid: string; d?: string };
-
 export type MessageListItemProps = {
 	message: IncompleteMessage;
 	selected: boolean;
