+++ src/views/app/folder-panel/messages/message-list-item.tsx
@@ -17,7 +17,7 @@ import { useMsgPreviewOnSeparatedWindowFn } from 'hooks/actions/use-msg-preview-
 import { useMsgSetReadFn } from 'hooks/actions/use-msg-set-read';
 import { useMarkAsReadOnClick } from 'hooks/use-mark-as-read-on-click';
 import { useOnMouseHover } from 'hooks/use-on-mouse-hover';
-import { MessageListItemProps } from 'types/index.d';
+import { MessageListItemProps } from 'types/folder';
 import { MessageListItemActionWrapper } from 'views/app/folder-panel/messages/message-list-item-action-wrapper';
 import { MessageListItemCore } from 'views/app/folder-panel/messages/message-list-item-core';
